pyproject.toml
[tool.poetry]
name = "github-stars-export"
version = "0.1.0"
description = ""
authors = ["Luc Shelton <lucshelton@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
click = "^8.1.7"
rich-click = "^1.8.3"
rich = "^13.8.1"
requests = "^2.32.3"
notion-client = "^2.2.1"
python-dotenv = "^1.0.1"

[tool.poetry.scripts]
github-stars-export = 'github_stars_export.__main__:cli'

[tool.black]
line-length = 100
target-version = ['py311']
include = '\.pyi?$'
# 'extend-exclude' excludes files or directories in addition to the defaults
extend-exclude = '''
# A regex preceded with ^/ will apply only to files and directories
# in the root of the project.
(
  ^/foo.py    # exclude a file named foo.py in the root of the project
  | .*_pb2.py  # exclude autogenerated Protocol Buffer files anywhere in the project
)
'''

[tool.poetry.group.docs.dependencies]
sphinx = "^7.4.4"
sphinx-rtd-theme = "^2.0.0"


[tool.poetry.group.dev.dependencies]
black = "^24.4.2"
pylint = "^3.2.5"


[tool.poetry.group.test.dependencies]
coverage = "^7.6.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"